analytics/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==42.0.4
    # via jwcrypto
django==5.0.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-oauth-toolkit
    #   djangorestframework
django-behaviors==0.5.1
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-oauth-toolkit==2.3.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
idna==3.6
    # via requests
jwcrypto==1.5.4
    # via django-oauth-toolkit
oauthlib==3.2.2
    # via django-oauth-toolkit
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.21
    # via cffi
pytz==2024.1
    # via djangorestframework
pyyaml==6.0.1
    # via -r requirements.in
requests==2.31.0
    # via django-oauth-toolkit
sqlparse==0.4.4
    # via django
typing-extensions==4.9.0
    # via jwcrypto
uritemplate==4.1.1
    # via -r requirements.in
urllib3==2.2.1
    # via requests
whitenoise==6.6.0
    # via -r requirements.in
